src/dataviewer/pyproject.toml
[project]
name = "datasetanalysistool"
version = "0.1.0"
description = "Web-based dataset analysis and annotation tool for robot demonstration datasets"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "python-multipart>=0.0.18",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.1.0",
    "numpy>=2.2.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.6.0",
    "pandas>=2.2.0",
    "pyarrow>=18.1.0",
    "Pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]
azure = [
    "azure-storage-blob>=12.24.0",
    "azure-identity>=1.19.0",
]
huggingface = [
    "huggingface-hub>=0.27.0",
]
hdf5 = [
    "h5py>=3.12.0",
]
export = [
    "h5py>=3.12.0",
    "numpy>=2.2.0",
    "Pillow>=10.0.0",
]
yolo = [
    "ultralytics>=8.3.0",
    "numpy>=2.2.0",
    "Pillow>=10.0.0",
]
all = [
    "datasetanalysistool[dev,azure,huggingface,hdf5,export]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008", "B905", "B904"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
